refactor: move `guildRank` to seperate file
